requirements.txt
phidata
packaging
yfinance
python-dotenv
duckduckgo-search
fastapi
uvicorn
groq
openai